improved messaging start/stop

+++ stop.sh
@@ -1,9 +1,26 @@
 #!/bin/sh
 
+# Checking if script is being run as root
+if [ $(id -u) -ne 0 ]; then
+  echo
+  echo "***********************************"
+  echo "* This script must be run as root *"
+  echo "***********************************"
+  echo
+  echo "Please issue 'sudo $0'"
+  echo
+  exit 0
+fi
+
 while read -r line
 do
   pid="$line"
+  echo "Killing process $pid.." 
   kill $pid
+  echo "-> DONE"
 done < "draco.pids.temp"
 
-echo "DrACO has been effectively shutdown". 
+# Printing result message
+echo "******************************"
+echo "*  DrACO has been shutdown   *"
+echo "******************************"
